requirements.txt
aiosqlite==0.19.0
databases==0.7.0
Faker==19.3.0
fastapi==0.101.1
greenlet==2.0.2
Jinja2==3.1.2
passlib==1.7.4
Pillow==10.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-multipart==0.0.6
SQLAlchemy==1.4.49
uvicorn==0.23.2
starlette==0.27.0